@@ -16,64 +16,64 @@
     "electron-start": "electron .",
     "local-start": "node local.js",
     "prebuild": "rm -rf dist/",
-    "build": "export PLATFORM=${PLATFORM:=darwin,linux,win32}; electron-packager . Awsaml --asar --ignore=node_modules/electron --ignore=test --ignore=dist --out=dist --platform=$PLATFORM --arch=x64 --app-version=${npm_package_version} --app-bundle-id=com.rapid7.awsaml --helper-bundle-id=com.rapid7.awsaml.helper",
+    "build": "export PLATFORM=${PLATFORM:=darwin,linux,win32}; node packager.js",
     "postbuild": "for platform in `echo $PLATFORM | sed 's/,/ /g'`; do export platform=$platform; npm run zip; done",
     "zip": "cd dist/Awsaml-${platform}-x64 && zip -q -y -FS -r ../awsaml-v${npm_package_version}-${platform}-x64.zip .",
     "test": "mocha",
     "cover": "rm -Rf coverage/* && istanbul cover _mocha -- -R spec",
-    "lint": "eslint *.js views/** lib/** test/**",
+    "lint": "eslint '*.js' 'views/**/*.jsx' 'lib/**/*.js' 'test/**/*.js'",
     "report": "coveralls < ./coverage/lcov.info",
     "watch": "webpack-dev-server",
     "pack": "webpack --progress --colors --config=webpack.config.pack.js",
     "show-appcast-checkpoint": "curl --compressed --location --user-agent 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_9_2) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/33.0.1750.152 Safari/537.36' 'https://github.com/rapid7/awsaml/releases.atom' | /usr/bin/sed 's|<pubDate>[^<]*</pubDate>||g' | shasum --algorithm 256"
   },
   "dependencies": {
-    "aws-sdk": "2.2.11",
-    "babel-preset-es2015": "^6.9.0",
-    "babel-preset-react": "^6.5.0",
-    "babel-preset-stage-0": "^6.5.0",
-    "body-parser": "1.18.2",
-    "cookie-parser": "1.4.0",
-    "electron": "1.7.9",
-    "electron-packager": "10.1.0",
+    "aws-sdk": "2.245.1",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-stage-0": "^6.24.1",
+    "body-parser": "1.18.3",
+    "cookie-parser": "1.4.3",
     "errorhandler": "1.5.0",
-    "express": "4.16.2",
-    "express-react-views": "^0.10.0",
-    "express-session": "1.13.0",
-    "ini": "1.3.4",
+    "express": "4.16.3",
+    "express-react-views": "^0.10.5",
+    "express-session": "1.15.6",
+    "ini": "1.3.5",
     "mkdirp": "0.5.1",
-    "morgan": "1.6.1",
-    "passport": "0.3.0",
-    "passport-saml": "0.31.0",
-    "react": "^15.1.0",
-    "react-dom": "^15.5.4",
-    "xmldom": "0.1.19",
-    "xpath.js": "1.0.6"
+    "morgan": "1.9.0",
+    "passport": "0.4.0",
+    "passport-saml": "0.33.0",
+    "prop-types": "^15.6.1",
+    "react": "^16.3.2",
+    "react-dom": "^16.3.2",
+    "xmldom": "0.1.27",
+    "xpath.js": "1.1.0"
   },
   "devDependencies": {
-    "babel-core": "^6.9.0",
-    "babel-eslint": "^6.0.4",
-    "babel-loader": "^6.2.4",
-    "babel-runtime": "^6.9.0",
-    "coveralls": "^2.11.9",
-    "css-loader": "^0.23.1",
-    "electron-rebuild": "^1.1.4",
-    "enzyme": "^2.3.0",
-    "eslint": "^2.10.2",
-    "eslint-config-rapid7": "^1.1.0",
-    "eslint-friendly-formatter": "^2.0.5",
-    "eslint-loader": "^1.3.0",
-    "eslint-plugin-import": "^1.8.0",
-    "eslint-plugin-jsx-a11y": "^1.2.2",
-    "eslint-plugin-react": "^5.1.1",
-    "istanbul": "^0.4.3",
-    "mocha": "^3.3.0",
-    "mocha-lcov-reporter": "^1.2.0",
-    "react-hot-loader": "^1.3.0",
-    "should": "^8.4.0",
-    "style-loader": "^0.13.1",
-    "typings": "^1.0.4",
-    "webpack": "^1.13.1",
-    "webpack-dev-server": "^1.14.1"
+    "babel-core": "^6.26.3",
+    "babel-eslint": "^8.2.3",
+    "babel-loader": "^7.1.4",
+    "babel-preset-env": "^1.7.0",
+    "babel-runtime": "^6.26.0",
+    "coveralls": "^3.0.1",
+    "css-loader": "^0.28.11",
+    "electron": "2.0.2",
+    "electron-packager": "12.1.0",
+    "electron-rebuild": "^1.7.3",
+    "enzyme": "^3.3.0",
+    "eslint": "^4.19.1",
+    "eslint-config-rapid7": "^2.11.1",
+    "eslint-friendly-formatter": "^4.0.1",
+    "eslint-loader": "^2.0.0",
+    "eslint-plugin-import": "^2.12.0",
+    "eslint-plugin-jsx-a11y": "^6.0.3",
+    "eslint-plugin-react": "^7.8.2",
+    "istanbul": "^0.4.5",
+    "mocha": "^5.2.0",
+    "mocha-lcov-reporter": "^1.3.0",
+    "react-hot-loader": "^4.2.0",
+    "should": "^13.2.1",
+    "style-loader": "^0.21.0",
+    "webpack": "^4.8.3",
+    "webpack-dev-server": "^3.1.4"
   }
 }
