@@ -1,5 +1,3 @@
-'use strict';
-
 const https = require('https');
 const express = require('express');
 const router = express.Router();
@@ -68,6 +66,7 @@ module.exports = (app, auth) => {
           metadataUrlValid: false,
           error: Errors.urlInvalidErr
         }));
+
         return;
       }
       Storage.set('metadataUrlValid', true);
