@@ -1,5 +1,3 @@
-'use strict';
-
 const express = require('express');
 const router = express.Router();
 
@@ -16,7 +14,9 @@ module.exports = (app) => {
     const sts = new Aws.STS();
     const session = req.session.passport;
 
-    if (session === undefined ) { return; }
+    if (session === undefined) {
+      return;
+    }
 
     const refreshResponseObj = Object.assign(ResponseObj, {
       accountId: session.accountId
@@ -30,6 +30,7 @@ module.exports = (app) => {
     }, (assumeRoleErr, data) => {
       if (assumeRoleErr) {
         res.redirect(config.auth.entryPoint);
+
         return;
       }
 
@@ -44,7 +45,7 @@ module.exports = (app) => {
       let metadataUrls = Storage.get('metadataUrls');
 
       // If the stored metadataUrl label value is the same as the URL default to the profile name!
-      if(metadataUrls[metadataUrl] === metadataUrl) {
+      if (metadataUrls[metadataUrl] === metadataUrl) {
         metadataUrls[metadataUrl] = profileName;
         Storage.set('metadataUrls', metadataUrls);
       }
