@@ -1,5 +1,3 @@
-'use strict';
-
 const fs = require('fs');
 
 class Storage {
@@ -21,12 +19,14 @@ class Storage {
     }
     if (!fs.existsSync(this.file)) {
       this.data = {};
+
       return;
     }
-    var json = fs.readFileSync(this.file, 'utf8');
+    const json = fs.readFileSync(this.file, 'utf8');
 
     if (json === '') {
       this.data = {};
+
       return;
     }
     this.data = JSON.parse(json);
@@ -64,8 +64,11 @@ class Storage {
     if (key in this.data) {
       value = this.data[key];
     }
+
     return value;
   }
 }
 
-module.exports = (path) => new Storage(path);
+module.exports = (path) => {
+  return new Storage(path);
+};
